refactor(playground): address review comments from axel7083 Signed-off-by: Marc Nuri <marc@marcnuri.com>
